Refactoring apply_tx_blob
