+++ module-system/sov-modules-stf-template/src/lib.rs
@@ -5,19 +5,21 @@ mod tx_verifier;
 
 pub use app_template::AppTemplate;
 pub use batch::Batch;
-use tracing::log::info;
+use tracing::{debug, error, log::info};
 pub use tx_verifier::RawTx;
 
 use sov_modules_api::{
     hooks::{ApplyBlobHooks, SyncHooks, TxHooks},
     Context, DispatchCall, Genesis, Spec,
 };
-use sov_rollup_interface::stf::StateTransitionFunction;
 use sov_rollup_interface::stf::{BatchReceipt, SyncReceipt};
+use sov_rollup_interface::stf::{StateTransitionFunction, TransactionReceipt};
 use sov_rollup_interface::zk::traits::Zkvm;
 use sov_state::StateCheckpoint;
 use sov_state::Storage;
 use std::io::Read;
+
+use crate::{app_template::ApplyBatchError, tx_verifier::TransactionAndRawHash};
 
 #[derive(Debug, Copy, Clone, PartialEq, Eq, serde::Serialize, serde::Deserialize)]
 pub enum TxEffect {
@@ -90,9 +92,153 @@ where
         blob: &mut impl sov_rollup_interface::da::BlobTransactionTrait,
         _misbehavior_hint: Option<Self::MisbehaviorProof>,
     ) -> BatchReceipt<Self::BatchReceiptContents, Self::TxReceiptContents> {
-        match self.apply_tx_blob(blob) {
-            Ok(batch) => batch,
-            Err(e) => e.into(),
+        debug!(
+            "Applying batch from sequencer: 0x{}",
+            hex::encode(blob.sender())
+        );
+
+        // Initialize batch workspace
+        let mut batch_workspace = self
+            .checkpoint
+            .take()
+            .expect("Working_set was initialized in begin_slot")
+            .to_revertable();
+
+        // ApplyBlobHook: begin
+        if let Err(e) = self.runtime.begin_blob_hook(blob, &mut batch_workspace) {
+            error!(
+                "Error: The batch was rejected by the 'begin_blob_hook' hook. Skipping batch without slashing the sequencer: {}",
+                e
+            );
+            // TODO: will be covered in https://github.com/Sovereign-Labs/sovereign-sdk/issues/421
+            self.checkpoint = Some(batch_workspace.revert());
+            return ApplyBatchError::Ignored(blob.hash()).into();
+        }
+        batch_workspace = batch_workspace.checkpoint().to_revertable();
+
+        // TODO: don't ignore these events: https://github.com/Sovereign-Labs/sovereign/issues/350
+        let _ = batch_workspace.take_events();
+
+        let (txs, messages) = match self.pre_process_batch(blob.data_mut()) {
+            Ok((txs, messages)) => (txs, messages),
+            Err(reason) => {
+                // Explicitly revert on slashing, even though nothing has changed in pre_process.
+                let mut batch_workspace = batch_workspace.revert().to_revertable();
+                let sequencer_da_address = blob.sender().as_ref().to_vec();
+                let sequencer_outcome = SequencerOutcome::Slashed {
+                    reason,
+                    sequencer_da_address: sequencer_da_address.clone(),
+                };
+                match self
+                    .runtime
+                    .end_blob_hook(sequencer_outcome, &mut batch_workspace)
+                {
+                    Ok(()) => {
+                        // TODO: will be covered in https://github.com/Sovereign-Labs/sovereign-sdk/issues/421
+                        self.checkpoint = Some(batch_workspace.checkpoint());
+                    }
+                    Err(e) => {
+                        error!("End blob hook failed: {}", e);
+                        self.checkpoint = Some(batch_workspace.revert());
+                    }
+                };
+
+                return (ApplyBatchError::Slashed {
+                    hash: blob.hash(),
+                    reason,
+                    sequencer_da_address,
+                })
+                .into();
+            }
+        };
+
+        // Sanity check after pre processing
+        assert_eq!(
+            txs.len(),
+            messages.len(),
+            "Error in preprocessing batch, there should be same number of txs and messages"
+        );
+
+        // Dispatching transactions
+        let mut tx_receipts = Vec::with_capacity(txs.len());
+        for (TransactionAndRawHash { tx, raw_tx_hash }, msg) in
+            txs.into_iter().zip(messages.into_iter())
+        {
+            // Pre dispatch hook
+            let sender_address = match self.runtime.pre_dispatch_tx_hook(&tx, &mut batch_workspace)
+            {
+                Ok(verified_tx) => verified_tx,
+                Err(e) => {
+                    // Don't revert any state changes made by the pre_dispatch_hook even if the Tx is rejected.
+                    // For example nonce for the relevant account is incremented.
+                    error!("Stateful verification error - the sequencer included an invalid transaction: {}", e);
+                    let receipt = TransactionReceipt {
+                        tx_hash: raw_tx_hash,
+                        body_to_save: None,
+                        events: batch_workspace.take_events(),
+                        receipt: TxEffect::Reverted,
+                    };
+
+                    tx_receipts.push(receipt);
+                    continue;
+                }
+            };
+            // Commit changes after pre_dispatch_tx_hook
+            batch_workspace = batch_workspace.checkpoint().to_revertable();
+
+            let ctx = C::new(sender_address.clone());
+            let tx_result = self.runtime.dispatch_call(msg, &mut batch_workspace, &ctx);
+
+            let events = batch_workspace.take_events();
+            let tx_effect = match tx_result {
+                Ok(_) => TxEffect::Successful,
+                Err(e) => {
+                    debug!(
+                        "Tx 0x{} was reverted error: {}",
+                        hex::encode(raw_tx_hash),
+                        e
+                    );
+                    // The transaction causing invalid state transition is reverted
+                    // but we don't slash and we continue processing remaining transactions.
+                    batch_workspace = batch_workspace.revert().to_revertable();
+                    TxEffect::Reverted
+                }
+            };
+            debug!("Tx {} effect: {:?}", hex::encode(raw_tx_hash), tx_effect);
+
+            let receipt = TransactionReceipt {
+                tx_hash: raw_tx_hash,
+                body_to_save: None,
+                events,
+                receipt: tx_effect,
+            };
+
+            tx_receipts.push(receipt);
+            // We commit after events have been extracted into receipt.
+            batch_workspace = batch_workspace.checkpoint().to_revertable();
+
+            // TODO: `panic` will be covered in https://github.com/Sovereign-Labs/sovereign-sdk/issues/421
+            self.runtime
+                .post_dispatch_tx_hook(&tx, &mut batch_workspace)
+                .expect("Impossible happened: error in post_dispatch_tx_hook");
+        }
+
+        // TODO: calculate the amount based of gas and fees
+        let sequencer_outcome = SequencerOutcome::Rewarded(0);
+
+        if let Err(e) = self
+            .runtime
+            .end_blob_hook(sequencer_outcome.clone(), &mut batch_workspace)
+        {
+            // TODO: will be covered in https://github.com/Sovereign-Labs/sovereign-sdk/issues/421
+            error!("Failed on `end_blob_hook`: {}", e);
+        };
+
+        self.checkpoint = Some(batch_workspace.checkpoint());
+        BatchReceipt {
+            batch_hash: blob.hash(),
+            tx_receipts,
+            inner: sequencer_outcome,
         }
     }
 
@@ -137,7 +283,7 @@ where
         match decoded {
             Ok(call) => {
                 // TODO: do something with this result
-                let _ =
+                let res =
                     self.runtime
                         .dispatch_call(call, &mut batch_workspace, &Context::new(address));
             }
