+++ module-system/sov-modules-stf-template/src/sync_strategies/mod.rs
@@ -1,2 +1,1 @@
 pub mod optimistic;
-pub mod zk;
